+++ lang/en/layouts.php
@@ -5,15 +5,14 @@ return [
     'modal.link_text' => 'Feedback',
     'modal.success_message' => 'Thank you for your feedback!',
 
-    'form.subject' => 'Subject',
-    'form.message' => 'Message',
-    'form.email' => 'Your Email',
+    'form.message' => 'Describe your issue',
+    'form.email' => 'Leave your email for contact (optional)',
     'form.submit' => 'Send Feedback',
 
-    'email.subject' => 'New Feedback from :app_name',
+    'email.subject' => 'Feedback :app_www',
     'email.no_subject' => 'No subject',
     'email.no_message' => 'No message',
     'email.from' => 'From',
-    'email.anonymous' => 'Anonymous',
+    'email.anonymous' => '-',
     'email.send_from' => 'Send from',
 ];