+++ resources/views/emails/feedback-plain.blade.php
@@ -4,10 +4,6 @@
 @lang('ig-common::messages.email.hello')
 
 
-@lang('ig-feedback::layouts.form.subject')
-
-    {{ $feedback['subject'] ?? __('ig-feedback::layouts.email.no_subject') }}
-
 @lang('ig-feedback::layouts.form.message')
 
     {{ $feedback['message'] ?? __('ig-feedback::layouts.email.no_message') }}
@@ -17,3 +13,9 @@
     {{ $feedback['email'] ?? __('ig-feedback::layouts.email.anonymous') }}
 
 @endsection
+
+@section('footer')
+@lang('ig-common::messages.email.regards'),
+{{ \InternetGuru\LaravelCommon\Support\Helpers::getAppInfo() }}
+{{ url()->previous() }}
+@stop